requirements.txt
fastapi
uvicorn
jinja2
pydantic
pyjwt
bcrypt
